# Livraria/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Livraria",
    "description": "Cadastro de Livrios",
    "contact": {
      "name": "Gabriel Brito Alves",
      "url": "https://www.linkedin.com/in/gabriel-brito-alves-7b8a2b188/",
      "email": "g.britoalves@yahoo.com.br"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Editoras": {
      "get": {
        "tags": [
          "Editoras"
        ],
        "summary": "Método para obter todos os cadastros disponíveis",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Editora"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Editoras"
        ],
        "summary": "Método para publicar um cadastro",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Informaçoes da editora para cadastro",
            "schema": {
              "$ref": "#/definitions/Editora"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Editora"
            }
          }
        }
      }
    },
    "/api/Editoras/{id}": {
      "get": {
        "tags": [
          "Editoras"
        ],
        "summary": "Método para obter um valor específico",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Id do valor que será obtido",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Editora"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Editoras"
        ],
        "summary": "Método para alterar um cadastro já salvo",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Id do cadastros",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Informaçoes do carastro para alteração",
            "schema": {
              "$ref": "#/definitions/Editora"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Editoras"
        ],
        "summary": "Método para deletar cadastro",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Id do cadastros que será deletado",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Editora"
            }
          }
        }
      }
    },
    "/api/Livros": {
      "get": {
        "tags": [
          "Livros"
        ],
        "summary": "Método para obter todos os cadastros disponíveis",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Livro"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Livros"
        ],
        "summary": "Método para publicar um cadastro",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Livro"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Livro"
            }
          }
        }
      }
    },
    "/api/Livros/{id}": {
      "get": {
        "tags": [
          "Livros"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Livro"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Livros"
        ],
        "summary": "Método para alterar um cadastro já salvo",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Id do cadastros",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Informaçoes do carastro para alteração",
            "schema": {
              "$ref": "#/definitions/Livro"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Livros"
        ],
        "summary": "Método para deletar cadastro",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Id do cadastros que será deletado",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Livro"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Editora": {
      "required": [
        "fundado",
        "nome",
        "sede"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "fundado": {
          "format": "int32",
          "type": "integer"
        },
        "sede": {
          "type": "string"
        }
      }
    },
    "Livro": {
      "required": [
        "nome",
        "preco"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "descricao": {
          "type": "string"
        },
        "idAutor": {
          "format": "int32",
          "type": "integer"
        },
        "idEditora": {
          "format": "int32",
          "type": "integer"
        },
        "idCategoria": {
          "format": "int32",
          "type": "integer"
        },
        "preco": {
          "format": "double",
          "type": "number"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter 'Bearer' [space] and then your token in the text input below.\r\n\r\nExample: \"Bearer 12345abcdef\""
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}